fixed a missing factor of 2 for the averaged spin polarization
The averaged spin polarization vector P^\mu = S^\mu/S, where S^\mu is the
spin vector and S = 1/2 for Lambda. Previously, we were computing the spin
vector S^\mu. In order to compare with the experiment polarization, we
need of divide the S^\mu by the spin of Lambda.

I also fixed a factor of 1/2 in the vorticity terms in the new derivation.