impose traceless and transversality for the sigma^{\munu} on surface
The hydro outputs sigma^{\mu\nu} which may or may not satisfy the
traceless and transversality condition. We checked that adding such a
regulation on the sigma^{\mu\nu} does not modify the final spin
polarization by more than 1%. But it is good to have these conditions
imposed in the calculations.