update the sampler to regulate local temperature and pressure using pure HRG EOS
This option is currently implemented at the mu_B = 0 case. Resetting the
local temperature to the pure HRG EoS typically introduce 0.5% corrections.
Such a correction is essential to acheive numerical verfication for
the particle sampler algorithm to reproduce the local energy-momentum
tensor.

We also update the tests settings, including an ideal test and a viscous
test for bulk viscous pressure. We will add more tests for the shear
stress tensor soon.

We also compute the event averaged total energy and momentum vector from
the particle samplers. These numbers are handed when one wants to check
the energy-momentum conservation from the hydrodynamic hyper-surface to
particle samples.